# Codigos/Back-End/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API - Sistema de Controle de Fluxos Veiculares",
    "description": "API para gerenciamento de usuários, integradores, municípios, equipamentos e fluxos veiculares.",
    "version": "1.0.0",
    "contact": {
      "name": "DNIT",
      "url": "https://www.gov.br/dnit"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Usuario": {
        "type": "object",
        "required": ["login", "senha", "nome", "ativo"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "login": {
            "type": "string",
            "maxLength": 30,
            "example": "admin"
          },
          "senha": {
            "type": "string",
            "maxLength": 100,
            "example": "senha123"
          },
          "nome": {
            "type": "string",
            "maxLength": 100,
            "example": "Administrador do Sistema"
          },
          "ativo": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Municipio": {
        "type": "object",
        "required": ["codigo", "descricao", "uf"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "codigo": {
            "type": "integer",
            "example": 3106200
          },
          "descricao": {
            "type": "string",
            "maxLength": 100,
            "example": "Belo Horizonte"
          },
          "uf": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO"
            ],
            "example": "MG"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          },
          "details": {
            "type": "string",
            "example": "Detalhes do erro"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          }
        }
      }
    },
    "parameters": {
      "idPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID do recurso"
      },
...
    }
  },
  "paths": {
    "/usuarios": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Listar usuários",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/usuarios/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Verifica credenciais de login",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "example": "admin"
                  },
                  "senha": {
                    "type": "string",
                    "example": "senha123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário logado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Obter usuário por ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "put": {
        "tags": ["Usuários"],
        "summary": "Atualizar usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Usuários"],
        "summary": "Excluir usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário excluído com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
...
    "/municipios": {
      "get": {
        "tags": ["Municípios"],
        "summary": "Listar municípios",
        "parameters": [
          {
            "name": "uf",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por destrito(UF)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de municípios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Municipio"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Municípios"],
        "summary": "Criar município",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Municipio"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Município criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Municipio"
                }
              }
            }
          }
        }
      }
    },
    "/municipios/{id}": {
      "get": {
        "tags": ["Municípios"],
        "summary": "Obter município por ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Município encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Municipio"
                }
              }
            }
          },
          "404": {
            "description": "Município não encontrado"
          }
        }
      },
      "put": {
        "tags": ["Municípios"],
        "summary": "Atualizar município",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Municipio"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Município atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Municipio"
                }
              }
            }
          },
          "404": {
            "description": "Município não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Municípios"],
        "summary": "Excluir município",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Município excluído com sucesso"
          },
          "404": {
            "description": "Município não encontrado"
          }
        }
      }
    },
...
  }
}