# Codigos_Refatorados/Back-End/controllers/municipios.controller.js
const models = require("../models");
const Validator = require("fastest-validator");
const router = require("express").Router();

router.get("/", async (req, res) => {
  
  whereMunicipios = {};
  if (req.query.uf) whereMunicipios.uf = req.query.uf.toUpperCase();

  models.Municipio.findAll({
    where: { ...whereMunicipios },
    order: [["descricao", "ASC"]],
  })
    .then((result) => {
      res.status(200).json(result);
    })
    .catch((err) => {
      res.status(500).json({
        message: "Não foi possível obter os municipios",
        error: err,
      });
    });
});

// function gerarQuery(req) {
//   const objQuery = {where: {}}
//   return
// }

router.get("/:id", async (req, res) => {
  const id = req.params.id;
  models.Municipio.findByPk(id)
    .then((result) => {
      if (result) {
        res.status(200).json(result);
      } else {
        res.status(404).json({
          message: "Municipio não encontrado",
        });
      }
    })
    .catch((err) => {
      res.status(500).json({
        message: "Não foi possível obter o municipio",
        error: err,
      });
    });
});

router.post("/", async (req, res) => {
  
  const municipio = {
    codigo: req.body.codigo,
    descricao: req.body.descricao,
    uf: req.body.uf,
  };

  const schema = {
    codigo: { type: "number", optional: false },
    descricao: { type: "string", optional: false, max: "255" },
    uf: {
      type: "enum",
      values: [
        "AC",
        "AL",
        "AP",
        "AM",
        "BA",
        "CE",
        "DF",
        "ES",
        "GO",
        "MA",
        "MT",
        "MS",
        "MG",
        "PA",
        "PB",
        "PR",
        "PE",
        "PI",
        "RJ",
        "RN",
        "RS",
        "RO",
        "RR",
        "SC",
        "SP",
        "SE",
        "TO",
      ],
      optional: false,
    },
  };

  const v = new Validator();
  const validationResponse = v.validate(municipio, schema);

  if (validationResponse !== true) {
    return res.status(400).json({
      message: "falha na validação!",
      errors: validationResponse,
    });
  }

  models.Municipio.create(municipio).then((result) => {
    res
      .status(201)
      .json({
        message: "Municipio criado com sucesso",
        municipio: {
          id: result.id,
          ...municipio,
        },
      })
      .catch((err) => {
        res.status(500).json({
          message: "Não foi possível incluir o municipio",
          error: err,
        });
      });
  });
});

router.put("/:id", async (req, res) => {
  
  const id = req.params.id;
  const municipioUpdate = {
    codigo: req.body.codigo,
    descricao: req.body.descricao,
    uf: req.body.uf,
  };

  const schema = {
    codigo: { type: "number", optional: false },
    descricao: { type: "string", optional: false, max: "255" },
    uf: {
      type: "enum",
      values: [
        "AC",
        "AL",
        "AP",
        "AM",
        "BA",
        "CE",
        "DF",
        "ES",
        "GO",
        "MA",
        "MT",
        "MS",
        "MG",
        "PA",
        "PB",
        "PR",
        "PE",
        "PI",
        "RJ",
        "RN",
        "RS",
        "RO",
        "RR",
        "SC",
        "SP",
        "SE",
        "TO",
      ],
      optional: false,
    },
  };

  const v = new Validator();
  const validationResponse = v.validate(municipioUpdate, schema);

  if (validationResponse !== true) {
    return res.status(400).json({
      message: "falha na validação!",
      errors: validationResponse,
    });
  }

  models.Municipio.update(municipioUpdate, {
    where: {
      id: id,
    },
  })
    .then((result) => {
      res.status(200).json({
        message: "Municipio atualizado com sucesso",
        municipio: {
          id: id,
          ...municipioUpdate,
        },
      });
    })
    .catch((err) => {
      res.status(500).json({
        message: "Não foi possível atualizar o municipio",
        error: err,
      });
    });
});

router.delete("/:id", async (req, res) => {
  const id = req.params.id;

  models.Municipio.destroy({
    where: {
      id: id,
    },
  })
    .then((result) => {
      res.status(200).json({
        message: "Municipio removido com sucesso",
      });
    })
    .catch((err) => {
      res.status(500).json({
        message: "Não foi possível remover o municipio",
        error: err,
      });
    });
});

module.exports = router;